Rename SHA3 opcode to KECCAK256
