+++ src/Nethermind/Nethermind.Evm.Test/GasPriceExtractorTests.cs
@@ -1,4 +1,4 @@
-﻿//  Copyright (c) 2021 Demerzel Solutions Limited
+//  Copyright (c) 2021 Demerzel Solutions Limited
 //  This file is part of the Nethermind library.
 // 
 //  The Nethermind library is free software: you can redistribute it and/or modify
@@ -23,124 +23,123 @@ using Nethermind.Evm.Tracing;
 using Nethermind.Serialization.Rlp;
 using NUnit.Framework;
 
-namespace Nethermind.Evm.Test
+namespace Nethermind.Evm.Test;
+
+[Explicit("Failing on MacOS GitHub Actions with stack overflow")]
+[TestFixture]
+public class GasPriceExtractorTests : VirtualMachineTestsBase
 {
-    [Explicit("Failing on MacOS GitHub Actions with stack overflow")]
-    [TestFixture]
-    public class GasPriceExtractorTests : VirtualMachineTestsBase
+    protected override long BlockNumber => MainnetSpecProvider.IstanbulBlockNumber;
+
+    [Test]
+    public void Block_header_rlp_size_assumption_is_correct()
     {
-        protected override long BlockNumber => MainnetSpecProvider.IstanbulBlockNumber;
+        Rlp rlp = BuildHeader();
 
-        [Test]
-        public void Block_header_rlp_size_assumption_is_correct()
+        rlp.Bytes.Length.Should().BeLessThan(600);
+    }
+
+    [Test]
+    public void Intrinsic_gas_cost_assumption_is_correct()
+    {
+        Rlp rlp = BuildHeader();
+
+        Transaction tx = Build.A.Transaction.WithData(rlp.Bytes).TestObject;
+        long gasCost = IntrinsicGasCalculator.Calculate(tx, Spec);
+        gasCost.Should().BeLessThan(21000 + 9600);
+    }
+
+    [Test]
+    public void Keccak_gas_cost_assumption_is_correct()
+    {
+        Rlp rlp = BuildHeader();
+
+        Transaction tx = Build.A.Transaction.WithData(rlp.Bytes).TestObject;
+        long gasCost = IntrinsicGasCalculator.Calculate(tx, Spec);
+        gasCost.Should().BeLessThan(21000 + 9600);
+
+        byte[] bytecode =
+            Prepare.EvmCode
+                .PushData("0x0200")
+                .PushData(0)
+                .PushData(0)
+                .Op(Instruction.CALLDATACOPY)
+                .PushData("0x0200")
+                .PushData(0)
+                .Op(Instruction.KECCAK256)
+                .Done;
+
+        (Block block, Transaction transaction) = PrepareTx(
+            BlockNumber, 1000000, bytecode, rlp.Bytes, 0);
+
+        CallOutputTracer callOutputTracer = new();
+        _processor.Execute(transaction, block.Header, callOutputTracer);
+        long minorCostsEstimate = 100;
+        long keccakCostEstimate = 30 + 512 / 6;
+        callOutputTracer.GasSpent.Should().BeLessThan(21000 + 9600 + minorCostsEstimate + keccakCostEstimate);
+    }
+
+    [Test]
+    public void Blockhash_times_256_gas_cost_assumption_is_correct()
+    {
+        Rlp rlp = BuildHeader();
+        byte[] bytecode =
+            Prepare.EvmCode
+                .PushData(256)
+                .Op(Instruction.JUMPDEST)
+                .PushData(1)
+                .Op(Instruction.DUP2)
+                .Op(Instruction.SUB)
+                .Op(Instruction.DUP1)
+                .Op(Instruction.BLOCKHASH)
+                .Op(Instruction.POP)
+                .PushData(0)
+                .Op(Instruction.DUP2)
+                .Op(Instruction.GT)
+                .PushData(3)
+                .Op(Instruction.JUMPI)
+                .Op(Instruction.STOP)
+                .Done;
+
+        (Block block, Transaction transaction) = PrepareTx(
+            BlockNumber, 1000000, bytecode, rlp.Bytes, 0);
+
+        CallOutputTracer callOutputTracer = new();
+        _processor.Execute(transaction, block.Header, callOutputTracer);
+        callOutputTracer.GasSpent.Should().BeLessThan(21000 + 9600 + 20000);
+    }
+
+    [Test]
+    public void Blockhash_times_256_no_loop()
+    {
+        Rlp rlp = BuildHeader();
+        Prepare bytecodeBuilder = Prepare.EvmCode
+                .PushData(0)
+                .PushData(1);
+
+        for (int i = 0; i < 256; i++)
         {
-            Rlp rlp = BuildHeader();
-
-            rlp.Bytes.Length.Should().BeLessThan(600);
+            bytecodeBuilder.Op(Instruction.ADD)
+                .Op(Instruction.BLOCKHASH)
+                .PushData(1);
         }
 
-        [Test]
-        public void Intrinsic_gas_cost_assumption_is_correct()
-        {
-            Rlp rlp = BuildHeader();
+        byte[] bytecode = bytecodeBuilder.Done;
 
-            Transaction tx = Build.A.Transaction.WithData(rlp.Bytes).TestObject;
-            long gasCost = IntrinsicGasCalculator.Calculate(tx, Spec);
-            gasCost.Should().BeLessThan(21000 + 9600);
-        }
+        (Block block, Transaction transaction) = PrepareTx(
+            BlockNumber, 1000000, bytecode, rlp.Bytes, 0);
 
-        [Test]
-        public void Keccak_gas_cost_assumption_is_correct()
-        {
-            Rlp rlp = BuildHeader();
+        CallOutputTracer callOutputTracer = new();
+        _processor.Execute(transaction, block.Header, callOutputTracer);
+        callOutputTracer.GasSpent.Should().BeLessThan(21000 + 9600 + 20000);
+    }
 
-            Transaction tx = Build.A.Transaction.WithData(rlp.Bytes).TestObject;
-            long gasCost = IntrinsicGasCalculator.Calculate(tx, Spec);
-            gasCost.Should().BeLessThan(21000 + 9600);
-
-            byte[] bytecode =
-                Prepare.EvmCode
-                    .PushData("0x0200")
-                    .PushData(0)
-                    .PushData(0)
-                    .Op(Instruction.CALLDATACOPY)
-                    .PushData("0x0200")
-                    .PushData(0)
-                    .Op(Instruction.SHA3)
-                    .Done;
-
-            (Block block, Transaction transaction) = PrepareTx(
-                BlockNumber, 1000000, bytecode, rlp.Bytes, 0);
-
-            CallOutputTracer callOutputTracer = new();
-            _processor.Execute(transaction, block.Header, callOutputTracer);
-            long minorCostsEstimate = 100;
-            long keccakCostEstimate = 30 + 512 / 6;
-            callOutputTracer.GasSpent.Should().BeLessThan(21000 + 9600 + minorCostsEstimate + keccakCostEstimate);
-        }
-
-        [Test]
-        public void Blockhash_times_256_gas_cost_assumption_is_correct()
-        {
-            Rlp rlp = BuildHeader();
-            byte[] bytecode =
-                Prepare.EvmCode
-                    .PushData(256)
-                    .Op(Instruction.JUMPDEST)
-                    .PushData(1)
-                    .Op(Instruction.DUP2)
-                    .Op(Instruction.SUB)
-                    .Op(Instruction.DUP1)
-                    .Op(Instruction.BLOCKHASH)
-                    .Op(Instruction.POP)
-                    .PushData(0)
-                    .Op(Instruction.DUP2)
-                    .Op(Instruction.GT)
-                    .PushData(3)
-                    .Op(Instruction.JUMPI)
-                    .Op(Instruction.STOP)
-                    .Done;
-
-            (Block block, Transaction transaction) = PrepareTx(
-                BlockNumber, 1000000, bytecode, rlp.Bytes, 0);
-
-            CallOutputTracer callOutputTracer = new();
-            _processor.Execute(transaction, block.Header, callOutputTracer);
-            callOutputTracer.GasSpent.Should().BeLessThan(21000 + 9600 + 20000);
-        }
-
-        [Test]
-        public void Blockhash_times_256_no_loop()
-        {
-            Rlp rlp = BuildHeader();
-            Prepare bytecodeBuilder = Prepare.EvmCode
-                    .PushData(0)
-                    .PushData(1);
-
-            for (int i = 0; i < 256; i++)
-            {
-                bytecodeBuilder.Op(Instruction.ADD)
-                    .Op(Instruction.BLOCKHASH)
-                    .PushData(1);
-            }
-
-            byte[] bytecode = bytecodeBuilder.Done;
-
-            (Block block, Transaction transaction) = PrepareTx(
-                BlockNumber, 1000000, bytecode, rlp.Bytes, 0);
-
-            CallOutputTracer callOutputTracer = new();
-            _processor.Execute(transaction, block.Header, callOutputTracer);
-            callOutputTracer.GasSpent.Should().BeLessThan(21000 + 9600 + 20000);
-        }
-
-        private static Rlp BuildHeader()
-        {
-            HeaderDecoder decoder = new();
-            BlockHeader blockHeader = Build.A.BlockHeader
-                .WithBloom(new Bloom(Enumerable.Repeat((byte) 1, 256).ToArray())).TestObject;
-            Rlp rlp = decoder.Encode(blockHeader);
-            return rlp;
-        }
+    private static Rlp BuildHeader()
+    {
+        HeaderDecoder decoder = new();
+        BlockHeader blockHeader = Build.A.BlockHeader
+            .WithBloom(new Bloom(Enumerable.Repeat((byte) 1, 256).ToArray())).TestObject;
+        Rlp rlp = decoder.Encode(blockHeader);
+        return rlp;
     }
 }
